+++ src/cidr.cpp
@@ -1,37 +1,25 @@
 #include <QtCore/QDebugStateSaver>
 
 #include "cidr.h"
-
-const QString slashFormat = QStringLiteral("%1/%2");
+#include "ipwatchdebug.h"
 
 namespace Cidr {
-
-Value::Value(QString string, bool isValid) : m_string(string), m_isValid(isValid) {
-}
-
-Value::Value(bool isValid) : m_isValid(isValid) {
-}
-
-bool Value::isValid() const {
-    return m_isValid;
-}
-
-bool Value::operator==(const Value &b) const {
-    return m_string == b.m_string;
-}
-
-bool Value::operator!=(const Value &b) const {
-    return m_string != b.m_string;
-}
-
-QString Value::string() const {
-    return m_string;
-}
 
 QDebug operator<<(QDebug debug, const Value &v) {
     QDebugStateSaver saver(debug);
     debug.nospace() << v.string();
     return debug;
+}
+
+Value current(const QString &interfaceName_,
+              uint prefixLength,
+              const AbstractNetworkInfo &networkInfo) {
+    for (const auto &entry : networkInfo.addressEntries(interfaceName_)) {
+        if (entry.ip().isGlobal() && entry.ip().protocol() == QAbstractSocket::IPv6Protocol) {
+            return toCidr(prefixLength, entry.ip().toIPv6Address());
+        }
+    }
+    return Value();
 }
 
 Value toCidr(const int prefixLength, const Q_IPV6ADDR &val_) {
@@ -41,12 +29,4 @@ Value toCidr(const int prefixLength, const Q_IPV6ADDR &val_) {
     return Value(slashFormat.arg(asString).arg(prefixLength), !asString.isEmpty());
 }
 
-Value current(const QString &interfaceName_, const uint prefixLength) {
-    for (auto entry : QNetworkInterface::interfaceFromName(interfaceName_).addressEntries()) {
-        if (entry.ip().isGlobal() && entry.ip().protocol() == QAbstractSocket::IPv6Protocol) {
-            return toCidr(prefixLength, entry.ip().toIPv6Address());
-        }
-    }
-    return Value();
-}
 } // namespace Cidr