+++ src/cidr.h
@@ -3,15 +3,27 @@
 #include <QtCore/QString>
 #include <QtNetwork/QNetworkInterface>
 
+#include "constants.h"
+#include "networkinfo.h"
+
 namespace Cidr {
+
 class Value {
 public:
-    explicit Value(QString string, bool isValid = false);
-    Value(bool isValid = false);
-    bool isValid() const;
-    bool operator==(const Value &b) const;
-    bool operator!=(const Value &b) const;
-    QString string() const;
+    Value(bool isValid = false) : m_isValid(isValid) {};
+    explicit Value(QString string, bool isValid) : m_string(string), m_isValid(isValid) {};
+    bool isValid() const {
+        return m_isValid;
+    };
+    bool operator==(const Value &b) const {
+        return m_string == b.m_string;
+    };
+    bool operator!=(const Value &b) const {
+        return m_string != b.m_string;
+    };
+    QString string() const {
+        return m_string;
+    };
 
 private:
     QString m_string;
@@ -19,7 +31,9 @@ private:
 };
 
 QDebug operator<<(QDebug debug, const Cidr::Value &v);
+Value toCidr(int prefixLength, const Q_IPV6ADDR &val_);
+Value current(const QString &interfaceName_,
+              uint prefixLength,
+              const AbstractNetworkInfo &networkInfo);
 
-Value toCidr(const int prefixLength, const Q_IPV6ADDR &val_);
-Value current(const QString &interfaceName_, const uint prefixLength);
 } // namespace Cidr